改进了粉丝查找和关注组的相关功能，废除了关注组的部分方法

+++ mypage/src/main/java/com/mp/mypage/user/service/UserService.java
@@ -8,6 +8,8 @@ import com.mp.mypage.user.dao.UserGroupMapper;
 import com.mp.mypage.user.dao.UserLabelMapper;
 import com.mp.mypage.user.dto.UserGroupDTO;
 import com.mp.mypage.user.dto.UserIdolDTO;
+import com.mp.mypage.user.dto.UserInfoDTO;
+import com.mp.mypage.user.dto.UserSimpleInfoDTO;
 import com.mp.mypage.user.entity.UserFollower;
 import com.mp.mypage.user.entity.UserGroup;
 import com.mp.mypage.user.entity.UserLabel;
@@ -23,7 +25,7 @@ import java.util.List;
  * @author 刘鑫源
  * @time 2019/11/12
  * @lastUpdateMan 刘鑫源
- * @lastUpdateTime 2019/11/17
+ * @lastUpdateTime 2019/1/10
  * @version 1.0
  */
 @Service
@@ -131,6 +133,9 @@ public class UserService {
      * @return 结果信息
      */
     public Result removeUserGroup(long groupId){
+        List<Long> idols = userFollowerMapper.selectAllIdolByGroupId(groupId);
+        if(idols != null && idols.size() != 0)
+            return new Result(Constant.UNABLE_DELETE, "分组无法删除");
         if(userGroupMapper.deleteByPrimaryKey(groupId) != 0)
             return new Result(Constant.OPERATOR_SUCCESS, "分组删除成功");
         else
@@ -155,11 +160,14 @@ public class UserService {
      * @param destId 关注者id
      * @return 结果信息， 包含粉丝列表
      */
-    public Result getAllFollowersByUserId(long destId){
-        List<Long> followerIds = userFollowerMapper.selectAllFollowerByUserId(destId);
-        if(followerIds != null)
+    public Result getAllFollowersByUserId(long destId, int pageNum, int pageSize){
+        PageHelper.offsetPage(pageNum * pageSize, pageSize);
+        List<UserSimpleInfoDTO> followers = userFollowerMapper.selectAllFollowerByUserId(destId);
+        PageInfo pageInfo = new PageInfo<>(followers);
+        if(followers != null)
             return new Result(Constant.OPERATOR_SUCCESS, "关注列表获取成功")
-                .setAttribute(followerIds);
+                    .addAttribute("pages", pageInfo.getPages())
+                    .addAttribute("followers", followers);
         else
             return new Result(Constant.RESULT_EMPTY, "结果为空");
     }
@@ -183,6 +191,7 @@ public class UserService {
      * @param groupId 用户分组id
      * @return 结果信息， 包含关注列表
      */
+    @Deprecated
     public Result getAllIdolByGroupId(long groupId){
         List<Long> idolIds = userFollowerMapper.selectAllIdolByGroupId(groupId);
         if(idolIds != null)
